+++ src/map_hl.cpp
@@ -25,7 +25,7 @@ void map_highlight_t::clear()
 }
 
 /* create a new item for the cursor */
-void map_t::hl_cursor_draw(lpos_t pos, unsigned int radius)
+void map_t::hl_cursor_draw(lpos_t pos, float radius)
 {
   cursor.reset(canvas->circle_new(CANVAS_GROUP_DRAW, pos, radius,
                                   0, style->highlight.node_color));
